Another performance test and results from local execution.

+++ src/morfologik/util/PerformanceTimer.java
@@ -71,7 +71,7 @@ public final class PerformanceTimer {
     /*
      * 
      */
-    public double elemsPerSecond(int sequences) {
+    public long elemsPerSecond(int sequences) {
 	return (long) (sequences / perRoundTime);
     }
 }
